refactor(web): 출력 부분 분리
